requirements.txt
py-cord[speed]==2.5.0
python-dotenv==1.0.1